:recycle: Extracted shared store as feature
