+++ src/feature/store-shared/index.ts
@@ -1,0 +1,1 @@
+export * from "./component/SharedStoreProvider";
